Se hicieron cambios varios. Se agregó la vista nueva de items y se vinculó al botón.

+++ Modulo1_Administracion/Models/combos.cs
@@ -6,7 +6,7 @@ namespace Modulo1_Administracion.Models
         [Key]
         public int id_combo {  get; set; }
         public string? descripcion { get; set; }
-        public double? precio { get; set; }
+        public decimal? precio { get; set; }
         public string? imagen { get; set; }
         public int? id_estado { get; set; }
     }
