+++ Modulo1_Administracion/Models/items_menu.cs
@@ -7,7 +7,7 @@ namespace Modulo1_Administracion.Models
         public int id_item_menu { get; set; }
         public string? nombre { get; set; }
         public string? descripcion { get; set; }
-        public double? precio { get; set; }
+        public decimal? precio { get; set; }
         public string? imagen { get; set; }
         public int? id_estado { get; set; }
         public int? id_categoria { get; set; }
